+++ api/src/main/java/org/allaymc/api/container/interfaces/BrewingStandContainer.java
@@ -1,6 +1,9 @@
 package org.allaymc.api.container.interfaces;
 
+import org.allaymc.api.block.data.BlockFace;
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.interfaces.ItemAirStack;
+import org.allaymc.api.item.type.ItemTypes;
 import org.jetbrains.annotations.Range;
 
 /**
@@ -8,7 +11,7 @@ import org.jetbrains.annotations.Range;
  *
  * @author daoge_cmd
  */
-public interface BrewingStandContainer extends BlockContainer {
+public interface BrewingStandContainer extends BlockContainer, SidedContainer {
     /**
      * The constant representing the reagent slot index in the brewing stand container.
      */
@@ -73,5 +76,52 @@ public interface BrewingStandContainer extends BlockContainer {
      */
     default void setResult(@Range(from = 0, to = 2) int slot, ItemStack itemStack) {
         setItemStack(slot + 1, itemStack);
+    }
+
+    @Override
+    default int[] getAllowedInsertSlots(BlockFace side, ItemStack stack) {
+        if (side == BlockFace.UP) {
+            return isBrewingReagent(stack) ? new int[]{REAGENT_SLOT} : new int[0];
+        }
+        if (side == BlockFace.DOWN) {
+            return new int[0];
+        }
+        if (isBrewingFuel(stack)) {
+            return new int[]{FUEL_SLOT};
+        }
+        if (isBrewingBottle(stack)) {
+            return new int[]{1, 2, 3};
+        }
+        return new int[0];
+    }
+
+    @Override
+    default int[] getAllowedExtractSlots(BlockFace side) {
+        if (side == BlockFace.DOWN) {
+            return new int[]{1, 2, 3};
+        }
+        return new int[0];
+    }
+
+    private static boolean isBrewingFuel(ItemStack stack) {
+        if (stack == null || stack == ItemAirStack.AIR_STACK) {
+            return false;
+        }
+        return stack.getItemType() == ItemTypes.BLAZE_POWDER;
+    }
+
+    private static boolean isBrewingBottle(ItemStack stack) {
+        if (stack == null || stack == ItemAirStack.AIR_STACK) {
+            return false;
+        }
+        var type = stack.getItemType();
+        return type == ItemTypes.GLASS_BOTTLE ||
+                type == ItemTypes.POTION ||
+                type == ItemTypes.SPLASH_POTION ||
+                type == ItemTypes.LINGERING_POTION;
+    }
+
+    private static boolean isBrewingReagent(ItemStack stack) {
+        return !isBrewingFuel(stack) && !isBrewingBottle(stack);
     }
 }